+++ src/core/builder.ts
@@ -3,14 +3,16 @@
  */
 
 import * as gulp from 'gulp';
-import {BuildConfig, BuildFunction, BuildFunctionObject, Options, Plugin, Stream} from "./types";
+import {BuildConfig, BuildFunction, BuildFunctionObject, GulpStream, Options, Plugin, Stream} from "./types";
 import {GPlugin} from "./plugin";
-import {is, toPromise} from "../utils/utils";
+import {info, is, toPromise} from "../utils/utils";
 import {GReloader} from "./reloader";
+import * as filter from "gulp-filter";
 
 export class GBuilder {
   stream: Stream;
   streamQ: Stream[] = [];
+  promiseQ: Promise<any>[] = [];
   conf: BuildConfig = {buildName: ''};
   buildOptions: Options = {};
   moduleOptions: Options = {};
@@ -40,6 +42,7 @@ export class GBuilder {
     this.conf = conf;
     this.buildOptions = conf.buildOptions || {};
     this.moduleOptions = conf.moduleOptions || {};
+    const flushStream = this.conf.flushStream;
     let promise = undefined;
 
     // preBuild
@@ -47,7 +50,10 @@ export class GBuilder {
 
     // build
     await this.build();
-    if (this.conf.flushStream) await toPromise(this.stream);
+    if (flushStream) {
+      await Promise.all(this.promiseQ);
+      await toPromise(this.stream);
+    }
     if (this.conf.copy) {   // copy is the last part of the build
       promise = this.chain(GPlugin.copy, {targets: this.conf.copy});
       if (this.conf.flushStream) await promise;
@@ -96,6 +102,7 @@ export class GBuilder {
 
   popStream() {
     if (this.streamQ.length > 0) {
+      if (this.stream) this.promiseQ.push(toPromise(this.stream));  // back for flushing
       this.stream = this.streamQ.pop()
     }
     return this;
@@ -127,11 +134,11 @@ export class GBuilder {
   sourceMaps(options: Options = {}) {
     if (!this.buildOptions.sourceMap) return this;
 
-    const smOpts = options.sourcemaps || this.moduleOptions.sourcemaps || {};
-    if (options.init)
-      this.pipe(require('gulp-sourcemaps').init(smOpts.init));
+    let opts = Object.assign({}, this.moduleOptions.sourcemaps, options);
+    if (opts.init)
+      this.pipe(require('gulp-sourcemaps').init(opts.init));
     else
-      this.pipe(require('gulp-sourcemaps').write(smOpts.dest || '.', smOpts.write));
+      this.pipe(require('gulp-sourcemaps').write(opts.dest || '.', opts.write));
     return this;
   }
 
@@ -141,17 +148,67 @@ export class GBuilder {
     return this;
   }
 
+  debug(options: Options={}) {
+    let title = options.title ? options.title : '';
+    // title = `[debugPlugin${title ? ':' + title : ''}]`;
+    let opts = Object.assign({}, this.moduleOptions.debug, options, {title});
+    return this.pipe(require('gulp-debug')(opts));
+  }
+
+  filter(pattern: string | string[] | filter.FileFunction = ["**", "!**/*.map"], options: filter.Options = {}) {
+    let opts = Object.assign({}, this.moduleOptions.filter, options);
+    return this.pipe(require('gulp-filter')(pattern, opts))
+  }
+
+  rename(options: Options={}) {
+    const opts = Object.assign({}, this.moduleOptions.rename, options.rename || options);
+    return this.pipe(require('gulp-rename')(opts));
+  }
+
+  //--- Stream contents handling API: sourceMaps() should be called ---
+
+  concat(options: Options = {}) {
+    const outFile = options.outFile || this.conf.outFile;
+    if (!outFile) {
+      if (options.verbose) info('[concatPlugin] Missing conf.outFile. No output generated.');
+      return this;
+    }
+    let opts = Object.assign({}, this.moduleOptions.concat, options.concat);
+
+    return this.filter().pipe(require('gulp-concat')(outFile, opts.concat)).sourceMaps();
+  }
+
+  minifyCss() {
+    return this.filter().chain(GPlugin.cleancss)
+      .rename({extname: '.min.css'}).sourceMaps();
+  }
+
+  minifyJs() {
+    return this.filter().chain(GPlugin.uglify)
+      .rename({extname: '.min.js'}).sourceMaps();
+  }
+
 
   /**----------------------------------------------------------------
-   * Builder utility functions: Returns value can be anything
+   * Builder utility functions: Return values can be anything
    *----------------------------------------------------------------*/
   call(action: Plugin, ...args: any[]): void | Promise<any> {
     return action instanceof GPlugin
       ? (action as GPlugin).process(this, ...args)
       : action(this, ...args);
   }
-
-  toPromise(stream: Stream) {
-    return toPromise(stream || this.stream);
-  }
+  //
+  // toPromise(stream: Stream): Promise<Stream>{
+  //   return toPromise(stream || this.stream);
+  // }
+  //
+  // cloneStream(): Stream {
+  //   return this.stream ? require('gulp-clone')(this.stream) : undefined;
+  // }
+  //
+  // mergeStream(stream: Stream): Stream {
+  //   if (stream)
+  //     this.stream = this.stream ? this.stream.pipe(require('gulp-clone')()) : stream;
+  //   return this.stream;
+  // }
 }